extract multi transactions widget from coin consolidation wizard
This code can be reused for a batch signing/broadcastion tool, that can process a series of transaction files.
The logic is unchanged, except for some minor label  changes.
