+++ .pre-commit-config.yaml
@@ -6,7 +6,8 @@ files: |
         electroncash/keystore.py|
         electroncash_gui/qt/address_list.py|
         electroncash_gui/qt/consolidate_coins_dialog.py|
-        electroncash_gui/qt/exception_window.py
+        electroncash_gui/qt/exception_window.py|
+        electroncash_gui/qt/multi_transactions_dialog.py
     )$
 repos:
 -   repo: https://github.com/pycqa/isort
