+++ electroncash_gui/qt/consolidate_coins_dialog.py
@@ -1,6 +1,4 @@
-import json
 from enum import Enum
-from pathlib import Path
 from typing import Optional, Sequence
 
 from PyQt5 import QtCore, QtWidgets
@@ -14,6 +12,7 @@ from electroncash.consolidate import (
 from electroncash.constants import PROJECT_NAME, XEC
 from electroncash.transaction import Transaction
 from electroncash.wallet import Abstract_Wallet
+from electroncash_gui.qt.multi_transactions_dialog import MultiTransactionsWidget
 from electroncash_gui.qt.util import MessageBoxMixin
 
 
@@ -120,7 +119,6 @@ class ConsolidateCoinsWizard(QtWidgets.QWizard, MessageBoxMixin):
         self.address: Address = address
         self.wallet: Abstract_Wallet = wallet
         self.transactions: Sequence[Transaction] = []
-        self.main_window = main_window
 
         self.coins_page = CoinSelectionPage()
         self.addPage(self.coins_page)
@@ -128,11 +126,8 @@ class ConsolidateCoinsWizard(QtWidgets.QWizard, MessageBoxMixin):
         self.output_page = OutputsPage(address)
         self.addPage(self.output_page)
 
-        self.tx_page = TransactionsPage()
+        self.tx_page = TransactionsPage(wallet, main_window)
         self.addPage(self.tx_page)
-        self.tx_page.save_button.clicked.connect(self.on_save_clicked)
-        self.tx_page.sign_button.clicked.connect(self.on_sign_clicked)
-        self.tx_page.broadcast_button.clicked.connect(self.on_broadcast_clicked)
 
         self.currentIdChanged.connect(self.on_page_changed)
 
@@ -176,66 +171,6 @@ class ConsolidateCoinsWizard(QtWidgets.QWizard, MessageBoxMixin):
         self.transactions = transactions
         can_sign = self.wallet.can_sign(self.transactions[0])
         self.tx_page.set_unsigned_transactions(self.transactions, can_sign)
-
-    def on_save_clicked(self):
-        dir = QtWidgets.QFileDialog.getExistingDirectory(
-            self, "Select output directory for transaction files", str(Path.home())
-        )
-        if not dir:
-            return
-        for i, tx in enumerate(self.transactions):
-            name = (
-                f"signed_{i:03d}.txn" if tx.is_complete() else f"unsigned_{i:03d}.txn"
-            )
-            path = Path(dir) / name
-
-            tx_dict = tx.as_dict()
-            with open(path, "w+", encoding="utf-8") as f:
-                f.write(json.dumps(tx_dict, indent=4) + "\n")
-        QtWidgets.QMessageBox.information(
-            self, "Done saving", f"Saved {len(self.transactions)} files to {dir}"
-        )
-
-    def on_sign_clicked(self):
-        password = None
-        if self.wallet.has_password():
-            password = self.main_window.password_dialog(
-                "Enter your password to proceed"
-            )
-            if not password:
-                return
-
-        for tx in self.transactions:
-            self.wallet.sign_transaction(tx, password, use_cache=True)
-
-        QtWidgets.QMessageBox.information(
-            self,
-            "Done signing",
-            f"Signed {len(self.transactions)} transactions. Remember to save them!",
-        )
-
-        # Signing is done in a different thread, so delay the check for completeness
-        # until the signatures have been added to the transaction.
-        QtCore.QTimer.singleShot(
-            2000,
-            lambda: self.tx_page.broadcast_button.setEnabled(
-                self.transactions[-1].is_complete()
-            ),
-        )
-        self.tx_page.save_button.setText("Save (signed)")
-
-    def on_broadcast_clicked(self):
-        self.main_window.push_top_level_window(self)
-        try:
-            for tx in self.transactions:
-                self.main_window.broadcast_transaction(tx, None)
-        finally:
-            self.main_window.pop_top_level_window(self)
-        QtWidgets.QMessageBox.information(
-            self,
-            "Done broadcasting",
-            f"Broadcasted {len(self.transactions)} transactions.",
-        )
 
 
 class CoinSelectionPage(QtWidgets.QWizardPage):
@@ -389,7 +324,7 @@ class OutputsPage(QtWidgets.QWizardPage):
 
 
 class TransactionsPage(QtWidgets.QWizardPage):
-    def __init__(self, parent=None):
+    def __init__(self, wallet, main_window, parent=None):
         super().__init__(parent)
         self.status: TransactionsStatus = TransactionsStatus.NOT_STARTED
         self.setTitle("Transactions")
@@ -400,46 +335,13 @@ class TransactionsPage(QtWidgets.QWizardPage):
         self.status_label = QtWidgets.QLabel()
         layout.addWidget(self.status_label)
 
-        self.num_tx_label = QtWidgets.QLabel()
-        layout.addWidget(self.num_tx_label)
-        self.num_in_label = QtWidgets.QLabel()
-        layout.addWidget(self.num_in_label)
-        self.in_value_label = QtWidgets.QLabel()
-        layout.addWidget(self.in_value_label)
-        self.out_value_label = QtWidgets.QLabel()
-        layout.addWidget(self.out_value_label)
-        self.fees_label = QtWidgets.QLabel()
-        layout.addWidget(self.fees_label)
-        self.reset_labels()
-
-        layout.addStretch(1)
-
-        buttons_layout = QtWidgets.QHBoxLayout()
-        layout.addLayout(buttons_layout)
-
-        self.save_button = QtWidgets.QPushButton("Save (unsigned)")
-        buttons_layout.addWidget(self.save_button)
-
-        self.sign_button = QtWidgets.QPushButton("Sign")
-        buttons_layout.addWidget(self.sign_button)
-
-        self.broadcast_button = QtWidgets.QPushButton("Broadcast")
-        self.broadcast_button.setEnabled(False)
-        buttons_layout.addWidget(self.broadcast_button)
-
-    def reset_labels(self):
-        self.num_tx_label.setText("Number of transactions:")
-        self.num_in_label.setText("Number of inputs per tx:")
-        self.in_value_label.setText("Input value:")
-        self.out_value_label.setText("Output value:")
-        self.fees_label.setText("Fees:")
+        self.multi_tx_display = MultiTransactionsWidget(wallet, main_window)
+        layout.addWidget(self.multi_tx_display)
 
     def display_work_in_progress(self):
         """Disable buttons, inform the user about the ongoing computation"""
-        self.reset_labels()
-        self.save_button.setEnabled(False)
-        self.sign_button.setEnabled(False)
-        self.broadcast_button.setEnabled(False)
+        self.multi_tx_display.reset_labels()
+        self.multi_tx_display.disable_buttons()
         self.setCursor(QtCore.Qt.WaitCursor)
 
     def update_status(self, status: TransactionsStatus):
@@ -455,32 +357,16 @@ class TransactionsPage(QtWidgets.QWizardPage):
             self.completeChanged.emit()
 
     def update_progress(self, num_tx: int):
-        self.num_tx_label.setText(f"Number of transactions: <b>{num_tx}</b>")
+        self.multi_tx_display.set_displayed_number_of_transactions(num_tx)
 
     def set_unsigned_transactions(
         self, transactions: Sequence[Transaction], can_sign: bool
     ):
-        """Enable buttons, compute and display some information about transactions."""
         self.unsetCursor()
         if not transactions:
             self.update_status(TransactionsStatus.NO_RESULT)
             return
-        # Reset buttons when fresh unsigned transactions are set
-        self.save_button.setText("Save (unsigned)")
-        self.save_button.setEnabled(True)
-        self.sign_button.setEnabled(can_sign)
-        self.broadcast_button.setEnabled(False)
-
-        # Assume the first transactions has the maximum number of inputs
-        num_in = 0 if len(transactions) == 0 else len(transactions[0].inputs())
-        self.num_in_label.setText(f"Maximum number of inputs per tx: <b>{num_in}</b>")
-
-        in_value = sum([tx.input_value() for tx in transactions]) / 100
-        out_value = sum([tx.output_value() for tx in transactions]) / 100
-        fees = sum([tx.get_fee() for tx in transactions]) / 100
-        self.in_value_label.setText(f"Input value: <b>{in_value} {XEC}</b>")
-        self.out_value_label.setText(f"Output value: <b>{out_value} {XEC}</b>")
-        self.fees_label.setText(f"Fees: <b>{fees} {XEC}</b>")
+        self.multi_tx_display.set_transactions(transactions, can_sign)
 
     def isComplete(self) -> bool:
         return self.status == TransactionsStatus.FINISHED
